Solve the static version number by serving the HTML from the server
The old method had the downside that part of the embeded data (the
version number for instance) was dynamic, and thus embeded assets was
required either to be built with knowledge of the future commit number,
or be ammended afterwards in the commit, which would make the commit
number just wrong.

The solution here is to let the server be the sole source of truth for
the version number, and dynamically build the index page with the needed
variables at runtime.