Use a QueryLink component for internal navigation
The component use the History API to trigger the navigation instead of
listening to the global click events in the App component. This is safer
as the interaction is more limited. Sadly, we have to trigger the
popstate event by hand because just pushing on the history stack isn't
enough.